Fix updating company label is not working.
`Company.labels` is a m2m field. Previously, this field relies on
ApiUtil.update_instance() to update just like normal primitive fields.
Now, this function is explicitly only updating for primitive field, and
not one2one or m2m fields, that's why it's not working now. Implementing
the dedicated update func for m2m is the way to go, which is what's done
in this commit.